refactor: refactor my-household
